+++ ry.sh
@@ -1,2 +1,2 @@
 #!/usr/bin/env bash
-npm run build && npm pack && npm i -g gazeall-0.9.0.tgz
+npm run build && npm pack && npm i -g gazeall-0.10.0.tgz