+++ src/lib/utils.ts
@@ -21,7 +21,6 @@ export interface GazeallOptions {
   run?: string[];
   wait?: boolean;
   watch?: string | string[];
-  files?: string[];
 }
 
 const PACKAGE_JSON = readPackageJSONProperties();
@@ -32,7 +31,7 @@ const PACKAGE_JSON = readPackageJSONProperties();
 let child_procs: ChildProcess[] = [];
 
 /**
- * Parse watch list.
+ * Parse list of files, folders to watch.
  * @param {GazeallOptions} cmd - The program arguments from commander module.
  * @returns {string[]} - Lsit of file or folder names.
  */
@@ -60,15 +59,19 @@ function getWatchList(cmd: any): string[] {
  * @return {void}
  */
 export function watchAndRun(cmd: any): void {
-  //  console.log('debug [watchAndRun]> ', cmd); // !debug
+  console.log('debug [watchAndRun]> ', cmd); // !debug
 
   if (cmd.npmp || cmd.npms) {
-    const watch_list = cmd.files.splice(2);
-    cmd.watch = watch_list.length > 0 ? watch_list : '**/*';
-    //    console.log('debug [watchAndRun:npmp] 0> ', cmd.npmp); // !debug
-    //    console.log('debug [watchAndRun:npms] 0> ', cmd.npms); // !debug
-    //    console.log('debug [watchAndRun:watch] 0> ', cmd.watch); // !debug
+    console.log('debug [watchAndRun:cmd] 0> ', cmd); // !debug
+    // Fallback to args if no watch files provided.
+    if (!cmd.watch) {
+      cmd.watch = cmd.args?.length > 0 ? cmd.args : '**/*';
+    }
+    console.log('debug [watchAndRun:npmp] 0> ', cmd.npmp); // !debug
+    console.log('debug [watchAndRun:npms] 0> ', cmd.npms); // !debug
+    console.log('debug [watchAndRun:watch] 0> ', cmd.watch); // !debug
   } else if (cmd.run) {
+    console.log('debug [watchAndRun:cmd] 1> ', cmd); // !debug
     // Called as: gazeall --run "node main.js" OR
     // Called as: gazeall --run "node main.js --watch <watch...>"
     if (!cmd.watch) {
@@ -77,48 +80,70 @@ export function watchAndRun(cmd: any): void {
     } else {
       cmd.watch = getWatchList(cmd);
     }
-    //    console.log('debug [watchAndRun:run] 1> ', cmd.run); // !debug
-    //    console.log('debug [watchAndRun:watch] 1> ', cmd.watch); // !debug
-    //    console.log('debug [watchAndRun:cmd] 1> ', cmd); // !debug
+    console.log('debug [watchAndRun:run] 1> ', cmd.run); // !debug
+    console.log('debug [watchAndRun:watch] 1> ', cmd.watch); // !debug
     // Called with command and no watch files.
-  } else if (!cmd.run && (cmd.watch || (cmd.files && cmd.files.length > 0))) {
+  } else if (!cmd.run && (cmd.watch || (cmd.args && cmd.args.length > 0))) {
+    console.log('debug [watchAndRun:cmd] 2> ', cmd); // !debug
     // Called as: gazeall main.js
     //        or: gazeall main.js <watch...> OR
-    //    console.log('debug [watchAndRun:cmd] 2> ', cmd); // !debug
-    //    console.log('debug [watchAndRun:files] 2> ', cmd.files); // !debug
     // Called with no run command, only a node.js run file.
     // This is the shorthand to run the watch file using Node.js.
     // Default to watch all files in current and all sub-folders.
-    cmd.run = [`node ${cmd.files[0]}`];
+    cmd.run = [`node ${cmd.args[0]}`];
 
     if (!cmd.watch) {
       // Default to watch all file in current and all sub-folders.
-      let watch = cmd.files.splice(1);
-      cmd.watch = watch.length > 0 ? watch : ['**/*.js'];
+      let watch = cmd.args.splice(1);
+      cmd.watch = watch.length > 0 ? watch : ['**/*'];
     } else {
       cmd.watch = getWatchList(cmd);
     }
-    //    console.log('debug [watchAndRun:run] 2> ', cmd.run); // !debug
-    //    console.log('debug [watchAndRun:watch] 2> ', cmd.watch); // !debug
-  } else if (!cmd.run && !cmd.watch && (!cmd.files || cmd.files.length === 0)) {
-    //    console.log('debug [watchAndRun:cmd] 3> ', cmd); // !debug
+    console.log('debug [watchAndRun:args] 2> ', cmd.args); // !debug
+    console.log('debug [watchAndRun:run] 2> ', cmd.run); // !debug
+    console.log('debug [watchAndRun:watch] 2> ', cmd.watch); // !debug
+  } else if (!cmd.run && !cmd.watch && (!cmd.args || cmd.args.length === 0)) {
+    console.log('debug [watchAndRun:cmd] 3> ', cmd); // !debug
 
     // Called as: gazeall
     // Called with no run and watch files.
     // Try to read run filename from package.json
     // The field "main" will be used as the file to execute using Node.js.
-    cmd.run = [`node ${PACKAGE_JSON['main']}`];
+    const package_json_main = PACKAGE_JSON['main'];
+    if (!package_json_main) {
+      console.log(
+        colors.red(
+          '=> Error: No run file passed and field "main" is missing in package.json, please correct one.'
+        )
+      );
+      process.exit(1);
+    }
+    cmd.run = [`node ${package_json_main}`];
     cmd.watch = ['**/*.js'];
-    //    console.log('debug [watchAndRun:run] 3> ', cmd.run); // !debug
-    //    console.log('debug [watchAndRun:watch] 3> ', cmd.watch); // !debug
+    console.log('debug [watchAndRun:run] 3> ', cmd.run); // !debug
+    console.log('debug [watchAndRun:args] 3> ', cmd.args); // !debug
+    console.log('debug [watchAndRun:watch] 3> ', cmd.watch); // !debug
+    console.log('debug [watchAndRun:npm_main] 3> ', PACKAGE_JSON['main']); // !debug
+    try {
+      fs.statSync(package_json_main);
+    } catch (ex: any) {
+      console.log(
+        colors.red(
+          `=> Error: File ${package_json_main} declared in package.json not found.`
+        )
+      );
+      process.exit(8);
+    }
   } else {
     console.log(
       colors.red(
-        'Missing package.json file, unable to read program name to run using Node.js.'
+        '=> Error: Missing package.json file, unable to read program name to run using Node.js.'
       )
     );
-    console.log(colors.red('Failed to provide a command to execute.'));
-    process.exit(1);
+    console.log(
+      colors.red('=> Error: Failed to provide a command to execute.')
+    );
+    process.exit(2);
   }
 
   // Check if we run first or wait first.
@@ -170,9 +195,9 @@ function stopRunningProcess(
       if (show_message) {
         console.log(
           colors.red(
-            `Stopping process with pid[${proc.pid}], exit[${proc.exitCode}] - ${
-              proc.exitCode === 0 ? 'success' : 'failed'
-            }`
+            `=> Stopping process with pid[${proc.pid}], exit[${
+              proc.exitCode
+            }] - ${proc.exitCode === 0 ? 'success' : 'failed'}`
           )
         );
       }
@@ -184,7 +209,9 @@ function stopRunningProcess(
  * Stop all running launched processes.
  */
 export function StopLaunchedProcesses() {
-  console.log(colors.red('\nStopping all launched processes.'));
+  console.log(
+    colors.red(`\n=> Stopping ${child_procs.length} launched processes.`)
+  );
   stopRunningProcess(child_procs, true);
   child_procs = [];
 }
@@ -199,44 +226,42 @@ function run(cmd: GazeallOptions): void {
   if (cmd.npmp) {
     // Run NPM scripts in parallel.
     const run_list: string[] = cmd.npmp.split(/\s+/);
-    //    console.log('debug [npmp]>', run_list); // !debug
+    console.log('debug [npmp]>', run_list); // !debug
     run_list.forEach((script: string) => {
       const command = PACKAGE_JSON['scripts'][script];
+      const pid = runCommand(command, cmd?.halt || false);
       console.log(
         colors.blue(
-          `=> Executing script: [${command}] => watching '${cmd.watch}'`
+          `=> Running script [${pid}:${command}] + Watching '${cmd.watch}'`
         )
       );
-      runCommand(command, cmd?.halt || false);
     });
   } else if (cmd.npms) {
     // Run NPM scripts in sequence.
     const run_list: string[] = cmd.npms.split(/\s+/);
-    //    console.log('debug [npms]>', run_list); // !debug
+    console.log('debug [npms]>', run_list); // !debug
     run_list.forEach((script: string) => {
       const command = PACKAGE_JSON['scripts'][script];
       console.log(
-        colors.blue(
-          `=> Executing script: [${command}] => watching '${cmd.watch}'`
-        )
+        colors.blue(`=> Running script [${command}] + Watching '${cmd.watch}'`)
       );
       const t_execution = runCommandSync(command, cmd?.halt || false);
       console.log(
-        colors.grey(`[${command}] => Execution completed (${t_execution} ms).`)
+        colors.grey(`=> Process [${command}] completed (${t_execution} ms).`)
       );
     });
   } else if (cmd.run) {
     // Run User supplied command.
-    console.log(
-      colors.blue(`=> Executing: [${cmd.run}] => watching '${cmd.watch}'`)
-    );
     cmd.run.forEach((command) => {
-      runCommand(command, cmd?.halt || false);
+      const pid = runCommand(command, cmd?.halt || false);
+      console.log(
+        colors.blue(`=> Running [${pid}:${command}] + Watching '${cmd.watch}'`)
+      );
     });
   } else {
     // Should never get here.
     console.log(colors.red('=> Error: Something went wrong, exiting!'));
-    process.exit(1);
+    process.exit(3);
   }
 }
 
@@ -246,20 +271,28 @@ function run(cmd: GazeallOptions): void {
  * @param {err_halt} boolean - Determines gazeall respose on an error,
  *                              If false, then ignore the error,
  *                              If true, then exit gazeall.
- * @return {void}
+ * @return {number} - Process PID.
  */
-function runCommand(command: string, err_halt: boolean): void {
+function runCommand(command: string, err_halt: boolean): number | undefined {
   const args: string[] = command.split(/\s+/);
   const cmd: string = args.shift() || '';
   const proc: ChildProcess = spawn(cmd, args, {detached: true});
-  child_procs.push(proc);
-
   const t_start = performance.now();
   //  console.log('debug t0=', t_start); // !debug
 
+  child_procs.push(proc);
+
   if (proc.stdout) {
     proc.stdout.on('data', (data: Buffer) => {
-      process.stdout.write(`[${command}] => ${data.toString()}`);
+      // Note: an output might contain newlines, so we want to append proc to each line displayed.
+      data
+        .toString()
+        .split('\n')
+        .filter((v) => v !== '')
+        .map((v) =>
+          process.stdout.write(`[${proc?.pid}:${command}] => ${v}\n`)
+        );
+      // process.stdout.write(`[${proc?.pid}:${command}] => ${data.toString()}`);
     });
   }
 
@@ -269,9 +302,11 @@ function runCommand(command: string, err_halt: boolean): void {
       if (err_halt) {
         stopRunningProcess(child_procs);
         process.stderr.write(
-          colors.red(`[${command}] => Error: Execution terminating.\n`)
+          colors.red(
+            `=> Error! Process [${proc?.pid}:${command}] terminating.\n`
+          )
         );
-        process.exit(1);
+        process.exit(4);
       }
     });
   }
@@ -281,10 +316,13 @@ function runCommand(command: string, err_halt: boolean): void {
     //    console.log('debug t0==', t_start); // !debug
     console.log(
       colors.grey(
-        `[${command}] => Execution completed (${t_end - t_start} ms).`
+        `=> Process [${proc?.pid}:${command}] completed (${
+          t_end - t_start
+        } ms).`
       )
     );
   });
+  return proc?.pid;
 }
 
 /**
@@ -296,7 +334,7 @@ function runCommand(command: string, err_halt: boolean): void {
  * @return {void}
  */
 // function runNPMCommand(command: string, err_halt: boolean): void {
-//    console.log( "debug [runNPMCommand] > ", command );  // !debug
+// console.log('debug [runNPMCommand] > ', command); // !debug
 //   const proc: ChildProcess = exec(
 //     `npm run ${command}`,
 //     (err, stdout, stderr) => {
@@ -336,13 +374,19 @@ function runCommandSync(command: string, err_halt: boolean): number {
     const out: Buffer | String = execSync(command);
     t_end = performance.now();
     if (out) {
-      process.stdout.write(`[${command}] => ${out.toString()}`);
+      // Note: an output might contain newlines, so we want to append proc to each line displayed.
+      out
+        .toString()
+        .split('\n')
+        .filter((v) => v !== '')
+        .map((v) => process.stdout.write(`[${command}] => ${v}\n`));
+      // process.stdout.write(`[${command}] => ${out.toString()}`);
     }
   } catch (err: any) {
     displayErrorMessage(err);
     if (err_halt) {
       stopRunningProcess(child_procs);
-      process.exit(1);
+      process.exit(5);
     }
   }
   return t_end - t_start;
@@ -351,28 +395,39 @@ function runCommandSync(command: string, err_halt: boolean): number {
 /**
  * Read package.json file and return settings as object.
  *
- * @returns package.json object
+ * @returns package.json object.
  */
-function readPackageJSONProperties() {
-  const file = path.join(process.cwd(), 'package.json');
-  let stats = fs.statSync(file);
+function readPackageJSONProperties(): any {
+  try {
+    const file = path.join(process.cwd(), 'package.json');
+    let stats = fs.statSync(file);
 
-  if (stats.isFile()) {
-    const data = fs.readFileSync(file, 'utf8');
-    const package_json = JSON.parse(data);
+    if (stats.isFile()) {
+      const data = fs.readFileSync(file, 'utf8');
+      const package_json = JSON.parse(data);
 
-    if (!package_json['main'] || package_json['main'] === '') {
-      throw new Error('Field main is missing or empty in package.json');
+      if (!package_json['main'] || package_json['main'] === '') {
+        console.log(
+          colors.red(
+            '=> Warning! Field main is missing or empty in package.json'
+          )
+        );
+      }
+
+      try {
+        stats = fs.statSync(package_json['main']);
+      } catch (ex: any) {
+        console.log(
+          colors.red(
+            `=> Warning! File ${package_json['main']} declared in package.json not found.`
+          )
+        );
+      }
+      return package_json;
     }
-
-    stats = fs.statSync(package_json['main']);
-
-    if (!stats.isFile()) {
-      throw new Error(
-        `File ${package_json['main']} declared in package.json not found.`
-      );
-    }
-    return package_json;
+  } catch (ex: any) {
+    console.log(colors.red(`=> Error: File package.json not found.`));
+    process.exit(6);
   }
-  return {};
+  return undefined;
 }